+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/RemoveAzManagmentGroupDeploymentStack.cs
@@ -17,7 +17,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Components;
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels;
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStacks;
-    using Microsoft.Azure.Management.Resources.Models;
+    using Microsoft.Azure.Management.Resources.DeploymentStacks;
+    using Microsoft.Azure.Management.Resources.DeploymentStacks.Models;
     using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
     using System;
     using System.Management.Automation;
@@ -68,6 +69,10 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         [Parameter(Mandatory = false, HelpMessage = "Bypass errors for the stack being out of sync when running the operation. If the stack is out of sync and this parameter " +
             "is not set, the operation will fail. Only include this parameter if instructed to do so on a failed stack operation.")]
         public SwitchParameter BypassStackOutOfSyncError { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "The action to take on resources that do not support deletion when they are removed from the deployment stack. " +
+            "Possible values include: 'Fail' (default) and 'Detach'.")]
+        public PSResourcesWithoutDeleteSupport? ResourcesWithoutDeleteSupport { get; set; }
 
         #endregion
 
@@ -128,7 +133,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                             resourcesCleanupAction: shouldDeleteResources ? "delete" : "detach",
                             resourceGroupsCleanupAction: shouldDeleteResourceGroups ? "delete" : "detach",
                             managementGroupsCleanupAction: shouldDeleteManagementGroups ? "delete" : "detach",
-                            bypassStackOutOfSyncError: BypassStackOutOfSyncError.IsPresent
+                            bypassStackOutOfSyncError: BypassStackOutOfSyncError.IsPresent,
+                            resourcesWithoutDeleteSupport: ResourcesWithoutDeleteSupport?.ToString().ToLowerInvariant()
                         );
                         if (PassThru.IsPresent)
                         {
@@ -139,7 +145,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
             }
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                 {
                     throw new PSArgumentException(dex.Message + " : " + dex.Body.Error.Code + " : " + dex.Body.Error.Message);
                 }