+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/TestAzManagementGroupDeploymentStack.cs
@@ -15,7 +15,8 @@
 namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
 {
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStacks;
-    using Microsoft.Azure.Management.Resources.Models;
+    using Microsoft.Azure.Management.Resources.DeploymentStacks;
+    using Microsoft.Azure.Management.Resources.DeploymentStacks.Models;
     using System;
     using System.Management.Automation;
 
@@ -61,7 +62,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         denySettingsExcludedActions: DenySettingsExcludedAction,
                         denySettingsApplyToChildScopes: DenySettingsApplyToChildScopes.IsPresent,
                         tags: Tag,
-                        bypassStackOutOfSyncError: BypassStackOutOfSyncError.IsPresent
+                        bypassStackOutOfSyncError: BypassStackOutOfSyncError.IsPresent,
+                        resourcesWithoutDeleteSupport: ResourcesWithoutDeleteSupport?.ToString().ToLowerInvariant(),
+                        validationLevel: ValidationLevel?.ToString()
                 );
 
                 if (PassThru.IsPresent)
@@ -71,7 +74,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
             }
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                     throw new PSArgumentException(dex.Message + " : " + dex.Body.Error.Code + " : " + dex.Body.Error.Message);
                 else
                     WriteExceptionError(ex);