+++ src/Resources/ResourceManager/SdkClient/DeploymentStacksSdkClient.cs
@@ -26,8 +26,8 @@ using System.Management.Automation;
 using System.Linq;
 using Microsoft.Rest.Azure;
 using System.Threading.Tasks;
-using Microsoft.Azure.Management.Resources;
-using Microsoft.Azure.Management.Resources.Models;
+    using Microsoft.Azure.Management.Resources.DeploymentStacks;
+    using Microsoft.Azure.Management.Resources.DeploymentStacks.Models;
 using ProjectResources = Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties.Resources;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments;
 using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Extensions;
@@ -119,7 +119,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                 {
                     if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
                     {
@@ -154,7 +154,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                 {
                     if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
                     {
@@ -188,7 +188,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                 {
                     if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
                     {
@@ -223,7 +223,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                 {
                     if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
                     {
@@ -268,7 +268,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 if (!throwIfNotExists)
                     return null;
 
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                     throw new PSArgumentException(dex.Body.Error.Message);
 
                 throw ex;
@@ -286,7 +286,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
 
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                 {
                     if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
                     {
@@ -331,7 +331,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 if (!throwIfNotExists)
                     return null;
 
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                     throw new PSArgumentException(dex.Body.Error.Message);
 
                 throw ex;
@@ -348,7 +348,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             }
             catch (Exception ex)
             {
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                 {
                     if (dex.Response.StatusCode == System.Net.HttpStatusCode.NotFound)
                     {
@@ -393,7 +393,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 if (!throwIfNotExists)
                     return null;
 
-                if (ex is DeploymentStacksErrorException dex)
+                if (ex is ErrorResponseException dex)
                     throw new PSArgumentException(dex.Body.Error.Message);
 
                 throw ex;
@@ -418,7 +418,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             string[] denySettingsExcludedActions,
             bool denySettingsApplyToChildScopes,
             Hashtable tags,
-            bool bypassStackOutOfSyncError
+            bool bypassStackOutOfSyncError,
+            string resourcesWithoutDeleteSupport = null,
+            string validationLevel = null
             )
         {
             // Create Deployment stack deployment model:
@@ -440,7 +442,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 denySettingsExcludedActions,
                 denySettingsApplyToChildScopes,
                 tags,
-                bypassStackOutOfSyncError
+                bypassStackOutOfSyncError,
+                resourcesWithoutDeleteSupport,
+                validationLevel
                 );
 
             ValidateDeploymentStack(deploymentStackModel, deploymentStackName, DeploymentStackScope.ResourceGroup, resourceGroupName);
@@ -478,7 +482,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
            string[] denySettingsExcludedActions,
            bool denySettingsApplyToChildScopes,
            Hashtable tags,
-           bool bypassStackOutOfSyncError
+           bool bypassStackOutOfSyncError,
+           string resourcesWithoutDeleteSupport = null,
+           string validationLevel = null
            )
         {
             // Create Deployment stack deployment model:
@@ -500,7 +506,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 denySettingsExcludedActions,
                 denySettingsApplyToChildScopes,
                 tags,
-                bypassStackOutOfSyncError
+                bypassStackOutOfSyncError,
+                resourcesWithoutDeleteSupport,
+                validationLevel
                 );
 
             ValidateDeploymentStack(deploymentStackModel, deploymentStackName, DeploymentStackScope.ResourceGroup, resourceGroupName);
@@ -512,11 +520,12 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             string resourcesCleanupAction, 
             string resourceGroupsCleanupAction, 
             string managementGroupsCleanupAction,
-            bool bypassStackOutOfSyncError
+            bool bypassStackOutOfSyncError,
+            string resourcesWithoutDeleteSupport = null
         )
         {
             var deleteResponse = DeploymentStacksClient.DeploymentStacks
-                .DeleteAtResourceGroupWithHttpMessagesAsync(resourceGroupName, name, resourcesCleanupAction, resourceGroupsCleanupAction, managementGroupsCleanupAction, bypassStackOutOfSyncError)
+                .DeleteAtResourceGroupWithHttpMessagesAsync(resourceGroupName, name, resourcesCleanupAction, resourceGroupsCleanupAction, managementGroupsCleanupAction, unmanageActionResourcesWithoutDeleteSupport: resourcesWithoutDeleteSupport, bypassStackOutOfSyncError)
                 .GetAwaiter()
                 .GetResult();
 
@@ -535,10 +544,11 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             string resourcesCleanupAction, 
             string resourceGroupsCleanupAction, 
             string managementGroupsCleanupAction,
-            bool bypassStackOutOfSyncError
+            bool bypassStackOutOfSyncError,
+            string resourcesWithoutDeleteSupport = null
         )
         {
-            var deleteResponse = DeploymentStacksClient.DeploymentStacks.DeleteAtSubscriptionWithHttpMessagesAsync(name, resourcesCleanupAction, resourceGroupsCleanupAction, managementGroupsCleanupAction, bypassStackOutOfSyncError)
+            var deleteResponse = DeploymentStacksClient.DeploymentStacks.DeleteAtSubscriptionWithHttpMessagesAsync(name, resourcesCleanupAction, resourceGroupsCleanupAction, managementGroupsCleanupAction, unmanageActionResourcesWithoutDeleteSupport: resourcesWithoutDeleteSupport, bypassStackOutOfSyncError)
                 .GetAwaiter()
                 .GetResult();
 
@@ -571,7 +581,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             string[] denySettingsExcludedActions,
             bool denySettingsApplyToChildScopes,
             Hashtable tags,
-            bool bypassStackOutOfSyncError
+            bool bypassStackOutOfSyncError,
+            string resourcesWithoutDeleteSupport = null,
+            string validationLevel = null
         )
         {
             // Create Deployment stack deployment model:
@@ -593,7 +605,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 denySettingsExcludedActions,
                 denySettingsApplyToChildScopes,
                 tags,
-                bypassStackOutOfSyncError
+                bypassStackOutOfSyncError,
+                resourcesWithoutDeleteSupport,
+                validationLevel
                 );
 
             ValidateDeploymentStack(deploymentStackModel, deploymentStackName, DeploymentStackScope.Subscription);
@@ -632,7 +646,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
         string[] denySettingsExcludedActions,
         bool denySettingsApplyToChildScopes,
         Hashtable tags,
-        bool bypassStackOutOfSyncError
+        bool bypassStackOutOfSyncError,
+        string resourcesWithoutDeleteSupport = null,
+        string validationLevel = null
 )
         {
             // Create Deployment stack deployment model:
@@ -654,7 +670,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 denySettingsExcludedActions,
                 denySettingsApplyToChildScopes,
                 tags,
-                bypassStackOutOfSyncError
+                bypassStackOutOfSyncError,
+                resourcesWithoutDeleteSupport,
+                validationLevel
                 );
 
             ValidateDeploymentStack(deploymentStackModel, deploymentStackName, DeploymentStackScope.Subscription);
@@ -666,11 +684,12 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             string resourcesCleanupAction, 
             string resourceGroupsCleanupAction, 
             string managementGroupsCleanupAction,
-            bool bypassStackOutOfSyncError
+            bool bypassStackOutOfSyncError,
+            string resourcesWithoutDeleteSupport = null
         )
         {
             var deleteResponse = DeploymentStacksClient.DeploymentStacks
-                    .DeleteAtManagementGroupWithHttpMessagesAsync(managementGroupId, name, resourcesCleanupAction, resourceGroupsCleanupAction, managementGroupsCleanupAction, bypassStackOutOfSyncError)
+                    .DeleteAtManagementGroupWithHttpMessagesAsync(managementGroupId, name, resourcesCleanupAction, resourceGroupsCleanupAction, managementGroupsCleanupAction, unmanageActionResourcesWithoutDeleteSupport: resourcesWithoutDeleteSupport, bypassStackOutOfSyncError)
                     .GetAwaiter()
                     .GetResult();
 
@@ -704,7 +723,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             string[] denySettingsExcludedActions,
             bool denySettingsApplyToChildScopes,
             Hashtable tags,
-            bool bypassStackOutOfSyncError
+            bool bypassStackOutOfSyncError,
+            string resourcesWithoutDeleteSupport = null,
+            string validationLevel = null
         )
         {
             // Create Deployment stack deployment model:
@@ -726,7 +747,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 denySettingsExcludedActions,
                 denySettingsApplyToChildScopes,
                 tags,
-                bypassStackOutOfSyncError
+                bypassStackOutOfSyncError,
+                resourcesWithoutDeleteSupport,
+                validationLevel
                 );
 
             ValidateDeploymentStack(deploymentStackModel, deploymentStackName, DeploymentStackScope.ManagementGroup, managementGroupId);
@@ -769,7 +792,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
            string[] denySettingsExcludedActions,
            bool denySettingsApplyToChildScopes,
            Hashtable tags,
-           bool bypassStackOutOfSyncError
+           bool bypassStackOutOfSyncError,
+           string resourcesWithoutDeleteSupport = null,
+           string validationLevel = null
        )
         {
             // Create Deployment stack deployment model:
@@ -791,7 +816,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 denySettingsExcludedActions,
                 denySettingsApplyToChildScopes,
                 tags,
-                bypassStackOutOfSyncError
+                bypassStackOutOfSyncError,
+                resourcesWithoutDeleteSupport,
+                validationLevel
                 );
 
             ValidateDeploymentStack(deploymentStackModel, deploymentStackName, DeploymentStackScope.ManagementGroup, managementGroupId);
@@ -815,14 +842,17 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
            string[] denySettingsExcludedActions,
            bool denySettingsApplyToChildScopes,
            Hashtable tags,
-           bool bypassStackOutOfSyncError
+           bool bypassStackOutOfSyncError,
+           string resourcesWithoutDeleteSupport = null,
+           string validationLevel = null
        )
         {
             var actionOnUnmanage = new ActionOnUnmanage
             {
                 Resources = resourcesCleanupAction,
                 ResourceGroups = resourceGroupsCleanupAction,
-                ManagementGroups = managementGroupsCleanupAction
+                ManagementGroups = managementGroupsCleanupAction,
+                ResourcesWithoutDeleteSupport = resourcesWithoutDeleteSupport
             };
             var denySettings = new DenySettings
             {
@@ -840,7 +870,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 DeploymentScope = deploymentScope,
                 DenySettings = denySettings,
                 Tags = TagsHelper.ConvertToTagsDictionary(tags),
-                BypassStackOutOfSyncError = bypassStackOutOfSyncError
+                BypassStackOutOfSyncError = bypassStackOutOfSyncError,
+                ValidationLevel = validationLevel
             };
 
             // Evaulate Template:
@@ -864,11 +895,11 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 // that preserves DateTime values as string (DateParseHandling = DateParseHandling.None),
                 // plus other custom settings (see: JsonExtensions.JsonObjectTypeSerializer)
                 deploymentStackModel.Template =
-                    FileUtilities.DataStore.ReadFileAsStream(templateFile).FromJson<JObject>();
+                    FileUtilities.DataStore.ReadFileAsStream(templateFile).FromJson<JObject>().ToObject<Dictionary<string, object>>();
             }
             else
             {
-                deploymentStackModel.Template = templateObject.ToJToken();
+                deploymentStackModel.Template = templateObject.ToJToken().ToObject<Dictionary<string, object>>();
             }
 
             // Evaluate Template Parameters:
@@ -999,7 +1030,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             WriteVerbose("Starting DeploymentOperations polling");
             try
             {
-                ResourceManagerSdkClient.ProvisionDeploymentStatus(parameters, new Deployment());
+                ResourceManagerSdkClient.ProvisionDeploymentStatus(parameters, new Microsoft.Azure.Management.Resources.Models.Deployment());
             }
             catch (Exception)
             {
@@ -1120,9 +1151,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 return null;
             }
 
-            if (ex is DeploymentStacksErrorException)
+            if (ex is ErrorResponseException)
             {
-                var stackEx = ex as DeploymentStacksErrorException;
+                var stackEx = ex as ErrorResponseException;
                 return new ErrorDetail(stackEx.Body?.Error.Code, stackEx.Body?.Error.Message, stackEx.Body?.Error.Target, stackEx.Body?.Error.Details);
             }
             else if (ex is CloudException)